Add author content upload state.
